+++ packages/video_player/video_player_android/lib/src/messages.g.dart
@@ -252,7 +252,13 @@ class PlatformVideoViewCreationParams {
 }
 
 class CreationOptions {
-  CreationOptions({required this.uri, this.formatHint, required this.httpHeaders, this.userAgent});
+  CreationOptions({
+    required this.uri,
+    this.formatHint,
+    required this.httpHeaders,
+    this.userAgent,
+    required this.enableDecoderFallback,
+  });
 
   String uri;
 
@@ -262,8 +268,10 @@ class CreationOptions {
 
   String? userAgent;
 
+  bool enableDecoderFallback;
+
   List<Object?> _toList() {
-    return <Object?>[uri, formatHint, httpHeaders, userAgent];
+    return <Object?>[uri, formatHint, httpHeaders, userAgent, enableDecoderFallback];
   }
 
   Object encode() {
@@ -277,6 +285,7 @@ class CreationOptions {
       formatHint: result[1] as PlatformVideoFormat?,
       httpHeaders: (result[2] as Map<Object?, Object?>?)!.cast<String, String>(),
       userAgent: result[3] as String?,
+      enableDecoderFallback: result[4]! as bool,
     );
   }